Find closest exon to splice junction ends

Find closest exon to splice junction ends

Usage

closestExonToJunctions(
  spliceGRgene,
  exonsGR,
  flipNegativeStrand = TRUE,
  sampleColname = "sample_id",
  reportActualCoords = FALSE,
  spliceBuffer = NULL,
  verbose = FALSE,
  ...
)

Arguments

spliceGRgene

GRanges representing splice junctions

exonsGR

GRanges representing flattened exons per gene, as is produced by flattenExonsBy().

flipNegativeStrand

logical indicating whether to flip the orientation of features on the negative strand.

sampleColname

character value matching the colname that defines distinct sample identifier, for which junctions will be kept separate.

reportActualCoords

logical indicating whether to report genomic coordinates or transcriptome coordinates. (Work in progress.)

spliceBuffer

optional integer indicating the maximum distance from an exon in order for an exon to be assigned to the exon. When spliceBuffer=NULL the distance is ignored in value columns c("nameFrom","nameTo"). When spliceBuffer is supplied, and junction is greater than this distance, the value columns c("nameFrom","nameTo") will indicate the exon name appended to the distance.

verbose

logical indicating whether to print verbose output.

...

additional arguments are ignored.

Details

This function is used to annotate splice junction GRanges entries based upon the closest compatible stranded exon boundary.

This function should usually be called by spliceGR2junctionDF() and not called directly.

See also