Summarize detected transcript results

Summarize detected transcript results

Usage

detectedTxInfo(
  detectedTxL,
  Gene = NULL,
  Tx = NULL,
  Groups = NULL,
  detectedOnly = NULL,
  ...
)

Arguments

detectedTxL

list output from defineDetectedTx()

Gene

optional character vector of one or more genes of interest, used to find transcript_id entries in the detectedTxL input data. If not supplied, Tx is expected.

Tx

optional character vector used to subset summary data, usually intended to keep rows in a specific order for a given set of transcripts.

Groups

optional character vector used to subset the groups (colnames) returned for each data.frame.

detectedOnly

NULL or logical indicating whether to restrict Tx to detected transcripts, defined by detectedTxL$detectedTx. When:

* `detectedOnly=NULL` then `Tx` is used as-is.
* `detectedOnly=TRUE` and `Tx=NULL` then `Tx` uses all detected transcripts from `detectedTxL$detectedTx`.
* `detectedOnly=TRUE` and `Tx` is supplied, it is restricted to detected transcripts from `detectedTxL$detectedTx`.
* `detectedOnly=FALSE` and `Tx=NULL` then `Tx` uses all transcripts.
* `detectedOnly=FALSE` and `Tx` is supplied, it is used as-is.

...

additional arguments are ignored.

Value

list of data.frame objects, each containing one summary table of data used to support whether each transcript were called "detected."

Details

This function provides a simple summary of the results of defineDetectedTx(), typically for a given gene of interest.

By default, either Gene or Tx must be supplied, however to return data for all genes, use either detectedOnly=TRUE to return data only for detected transcripts, or detectedOnly=FALSE to return all data including transcripts that are not determined to be detected.

See also