Get gaps in GRangesList objects

Get gaps in GRangesList objects

Usage

getGRLgaps(grl, strandSpecific = TRUE, verbose = FALSE, ...)

Arguments

grl

GRangesList object.

strandSpecific

logical, default TRUE, whether to determine gaps within strand.

verbose

logical indicating whether to print verbose output.

...

additional arguments are ignored.

Value

GRangesList object with gaps for each chromosome and strand present in each GRanges entry. It does not return gap sequence at the edges of GRanges regions to the chromosome ends.

Details

This function returns gaps between GRanges regions in a GRangesList object. When strandSpecific=TRUE is determines gaps per strand, otherwise strands are converted to "*". It will also determine gaps within chromosome for each GRanges entry in GRangesList.

See also