maximum overlapping internal junction score

maximum overlapping internal junction score

Usage

internal_junc_score(
  juncGR,
  scoreColname = "score",
  sampleColname = "sample_id",
  minScore = 0,
  verbose = FALSE,
  ...
)

Arguments

juncGR

GRanges containing splice junctions, with numeric column scoreColname representing the abundance of splice junction-spanning sequence reads.

scoreColname, sampleColname

colnames in values(juncGR) which define the junction score, and the "sample_id".

minScore

optional numeric value indicating the minimum score to use, which can be useful for example if there is no matching scoreColname. Note that values are matched using absolute value, but the original sign is maintained.

verbose

logical indicating whether to print verbose output.

...

additional arguments are ignored.

Value

numeric vector named by names(juncGR) whose values are the maximum score of internal overlapping junction ends.

Details

This function is intended for internal use, and calculates the maximum score for junction GRanges for the special case where:

a junction range overlaps the start or end of another junction

the start or end of the other junction is internal, that is it does not overlap the same start or end.

overlaps are constrained to the same "sample_id" stored in values(juncGR)$sample_id.

The goal is to determine the highest score contained inside each junction region, so that the junction arc height can be defined higher in order to minimize junction arc overlaps.

See also