read gtf/gff3 file

read gtf/gff3 file

Usage

readGtf(GTF, nrows = -1, zcat_command = "zcat", verbose = FALSE, ...)

Arguments

GTF

character file name sent to data.table::fread(). When the file ends with ".gz", the R.utils package is recommended, otherwise the fallback option is to make a system call to gzcat to gunzip the file during the import step. Note this process fails when gzcat is not available in the path of the user environment. In general, the R.utils package is the best solution.

nrows

integer number of rows to read from the GTF file, by default -1 means all rows are imported. This parameter is useful to check the results of a large GTF file using only a subset portion of the file.

zcat_command

character name or path to zcat or gzcat executable, only used when input GTF is a file with ".gz" extension, and when R package R.utils is not available.

verbose

logical whether to print verbose output during processing.

...

additional arguments are ignored.

Value

data.frame representing tab-delimited data stored in the gtf or gff3 file.

See also