Assign exon names to GRangesList

Assign exon names to GRangesList

assignGRLexonNames(
  GRL,
  geneSymbolColname = "geneSymbol",
  exonNameColname = paste0(geneSymbolColname, "Exon"),
  suffix = "_exon",
  renameOnes = FALSE,
  filterTwoStrand = TRUE,
  checkDisjoin = c("warn", "none", "stop"),
  assignGRLnames = TRUE,
  verbose = FALSE,
  ...
)

Arguments

GRL
GRangesList input object. Ideally, the input GRanges are disjoint, meaning no two exons overlap, but instead are represented by non-overlapping regions that are sometimes adjacent.

geneSymbolColname
character string indicating with colname of values(GRL) contains the unique gene symbol. If this value does not already exist, it is created and populated using names(GRL).

exonNameColname
character string to be used for the resulting exon name. By default "Exon" is appended to the end of the geneSymbolColname.

suffix
character value indicating the suffix to add to the geneSymbolColname to indicate individual exon numbers.

renameOnes
logical indicating whether to name exon sub-sections when the exon is not subdivided. For example, when renameOnes=FALSE, an exon with one section would be named, "exon1", otherwise when renameOnes=TRUE an exon with one section would be named, "exon1a". This distinction can be helpful to recognize exons that contain no subsections.

filterTwoStrand
logical indicating whether to filter out genes occurring on two different strands.

checkDisjoin
character value indicating how to handle non-disjoint input GRL ranges. When checkDisjoin="stop" then any non-disjoint GRanges in an element of GRL will cause the function to fail.

assignGRLnames
logical indicating whether names for the resulting GRangesList should use the exon names.

verbose
logical indicating whether to print verbose output.

...
additional arguments are ignored.

Value

GRangesList object with additional columns indicating the exon name.

Details

This function takes a GRangesList object with an annotated gene symbol column, and defines exon numbers for each distinct (non-adjacent) range per gene. When multiple ranges overlap, one exon number is applied to them all, and disjoint ranges are denoted using a letter suffix.

For example the exon labels below:

|======|......|======|=======|======|......|======|=======| .exon1.........exon2a.exon2b..exon2c........exon3a..exon3b.

The full name for each feature will become:

Gene_exon1

Gene_exon2a

Gene_exon2b

Gene_exon2c

Gene_exon3a

Gene_exon3b

The reasoning, is to preserve the gene symbol for readability, but to number exons to indicate the numbered contiguous exon, with suffix to indicate the sub-section of each exon.

See also