Find overlaps between two GRangesList objects

Find overlaps between two GRangesList objects

findOverlapsGRL(GRL1, GRL2, annoName1 = "name", annoName2 = "name",
  check_names = TRUE, ...)

Arguments

GRL1
GRangesList query

GRL2
GRangesList subject

annoName1
character value indicating either the colname of values(GRL1) to use as the name, or if "name" then it uses names(GRL1).

annoName2
character value indicating either the colname of values(GRL2) to use as the name, or if "name" then it uses names(GRL2).

check_names
logical indicating whether the values defined by annoName1 should match the values defined by annoName2. Note that when check_names=FALSE the overlaps returned will depend upon GRL1 and GRL2 being in identical order. Otherwise, the values in annoName1 and annoName2 are used, which means GRL1 and GRL2 do not have to be in identical order.

...
additional arguments are passed to GenomicRanges::findOverlaps(), useful for customizing the overlap criteria.

Value

Hits object, or the natural output from GenomicRanges::findOverlaps() dependent upon the ... arguments, subsetted for only entries with matching values defined by annoName1 and annoName2.

Details

This function implements GenomicRanges::findOverlaps() for the special case of two GRangesList objects, restricting results to those including the same GRangesList index in the subject and query.

See also