Flatten exons by gene or transcript

Flatten exons by gene or transcript

flattenExonsBy(exonsByTx, tx2geneDF, by = c("gene", "tx"),
  detectedTx = NULL, genes = NULL, txColname = "transcript_id",
  geneColname = "gene_name", cdsByTx = NULL, cdsByGene = NULL,
  verbose = FALSE)

Arguments

exonsByTx
GRangesList named by transcript, containing one or more GRanges representing exons. This data is often produced from TxDb data using GenomicFeatures::exonsBy(...,by="tx").

tx2geneDF
data.frame containing at least two columns with transcript and gene annotation, whose colnames are defined by arguments txColname and geneColname respectively. When using a GTF file, makeTx2geneFromGtf() can be used to create a tx2geneDF in data.frame format.

by
character string to group exons, "gene" groups multiple transcripts per gene, and "tx" groups exons per transcript. Note that in both cases, it combines exonsByTx and cdsByTx when cdsByTx is also supplied.

detectedTx
character vector of detected transcripts, used to subset the overall transcripts prior to producing a flattened gene exon model.

genes
optional character vector, representing a subset of genes for which flattened exons will be prepared. This argument is useful when focusing on only one or a subset of genes.

txColname
character string indicating a column from colnames(tx2geneDF) used to identify transcripts.

geneColname
character string indicating a column from colnames(tx2geneDF) used to identify gene name, or gene symbol.

cdsByTx
GRangesList named by transcript, containing GRanges exons that only include CDS regions. This data is often produced from TxDb data using GenomicFeatures::cdsBy(...,by="tx").

cdsByGene
GRangesList named by gene, containing GRanges exons that only include CDS regions. This data is often produced from TxDb data using GenomicFeatures::cdsBy(...,by="gene"). Note this input is only used when by="gene".

verbose
logical indicating whether to print verbose output.

Value

GRangesList named by gene when by="gene" or transcript when by="tx", containing non-overlapping GRanges with exon names as described above.

Details

This function takes as input:

exonsByTx as a GRangesList object of transcript exons named by the transcript_id,

tx2geneDF a data.frame with transcript-gene cross-reference,

detectedTx an optional character vector of transcript_id values, used to subset the overall transcripts

cdsByTx an optional GRangesList object, similar to exonsByTx except that it only contains the CDS portion of exons

This function groups exons together by gene, producing a flattened, disjoint (non-overlapping) set of exons, where exons are subdivided when there are multiple boundaries.

Finally, it labels each exon using a defined naming scheme:

Each contiguous exon is numbered in order, starting at 1 for the first stranded exon for the gene, for example exon1, exon2, exon3.

When an exon is sub-divided, each section is labeled with an alphabetic suffix to indicate the order within that exon, for example exon1a, exon1b, exon1c.

A text schematic is shown below:

|=======|======|......|=======|.....|=======|=======|=======|

|.exon1a|exon1b|......|.exon2.|.....|.exon3a|.exon3b|.exon3c|

Where

|====| represents an exon,

|====|====| represents one contiguous exon with two sub-divided parts, and

..... represents an intron.

It is recommended but not required to supply detectedTx, since it can greatly reduce the total number of transcripts. This step has two benefits: It can greatly simplify the resulting exon model based upon observed data, and it has the by-product of removing potentially erroneous transcripts which often has no supporting observed data.

See also