Gene GRangesList to ggplot2 grob

Gene GRangesList to ggplot2 grob

gene2gg(gene = NULL, tx = NULL, flatExonsByGene = NULL,
  flatExonsByTx = NULL, geneColor = "dodgerblue", labelExons = TRUE,
  exonLabelAngle = 90, exonLabelSize = 3,
  newValues = list(feature_type = "gap", subclass = "gap", gene_nameExon
  = "gap"), gene_order = c("first", "last"), return_type = c("grob",
  "df"), ref2c = NULL, hjust = 1.2, vjust = 0, compressGaps = TRUE,
  tx2geneDF = NULL, verbose = FALSE, ...)

Arguments

gene
character string of the gene to plot, compared with names(flatExonsByGene) and values(flatExonsByTx)$gene_name.

tx
character vector of the transcripts to plot, useful when specifying specific transcripts. Values are matched with names(flatExonsByTx).

flatExonsByGene, flatExonsByTx
GRangesList objects, named by "gene_name" or "transcript_id" respectively, containing disjoint (non-overlapping) exons within each GRangesList element. The data is expected to be in the form provided by flattenExonsBy().

geneColor
character color used as the base color for exons, where the color is varied for each feature type or subclass.

labelExons
logical indicating whether to print text labels beneath each exon, using the values in colname "gene_nameExon". Typically the gene and transcripts are named using consistent names, in which case one exon label is placed at the bottom of the lowest transcript for each unique exon label.

exonLabelAngle
numeric angle in degrees (0 to 360) indicating how to rotate exon labels, where 90 is vertical, and 0 is horizontal.

exonLabelSize
numeric value, compatible with argument size in ggrepel::geom_text_repel(), used to size exon labels when labelExons=TRUE.

newValues
argument passed to addGRLgaps() to fill column values for newly created gap entries. It is useful to have feature_type="gap" so gaps have a different value than exons. It is also useful to have subclass="gap" when there are "cds" and "noncds" entries in the provided flatExonsByGene data.

gene_order
character value indicating whether the flattened gene model should be plotted "first" above the transcript exon models, or "last" and below the transcript exon models.

return_type
character value indicating whether to return the ggplot graphic object "grob", or the data.frame "df" used to create the ggplot object.

ref2c
list output from make_ref2compressed() which contains among other things, the trans_grc data of class "trans" used in ggplot2::coord_trans() or ggplot2::scale_x_continuous().

hjust, vjust
numeric value to position exon labels using ggplot::geom_text().

compressGaps
logical indicating whether to compress gaps between exons. When ref2c is supplied, this argument is ignored and the supplied ref2c is used directly.

tx2geneDF
data.frame or NULL, optionally used to help identify matching transcripts for the requested gene value, used when "gene_name" is not present in values(flatExonsByTx).

...
additional arguments are passed to relevant functions as needed, including make_ref2compressed().

Details

This function is intended to help plot gene and transcript exon models, and is a lightweight wrapper around grl2df().

It takes flatExonsByGene which is the output from flattenExonsBy(), and essentially plots the end result for review.

Alternatively, when return_type="df", the output is the data.frame used to produce the ggplot, which allows for more customization.

See also

Examples

## Assume we start with flattened gene exons
if (1 == 2) {
   ## Do not run automated examples until sample data is available

ggGria1 <- gene2gg("Gria1",
   flatExonsByGene=flatExonsByGeneCds);

## if transcript exons are available
ggGria1 <- gene2gg("Gria1",
   flatExonsByGene=flatExonsByGene,
   flatExonsByTx=flatExonsByTx);

}