Get coverage for GRanges from bigWig files

Get coverage for GRanges from bigWig files

getGRcoverageFromBw(gr, bwUrls, addGaps = FALSE,
  gap_feature_type = "gap", default_feature_type = "exon",
  feature_type_colname = "feature_type", verbose = FALSE, ...)

Arguments

gr
GRanges object

bwUrls
character vector of full file paths or web URLs to bigWig files, suitable for use by rtracklayer::import().

addGaps
logical indicating whether gaps between GRanges should be added to the query. Gaps are determined using getGRgaps().

feature_type_colname, gap_feature_type, default_feature_type
When addGaps=TRUE a new column named using feature_type_colname is added to values(gr), whose value for gap regions is gap_feature_type. When feature_type_colname is already present in gr it is not modified, otherwise the column is created with value default_feature_type. By default, this function adds a column "feature_type" with value "gap".

verbose
logical indicating whether to print verbose output.

...
additional arguments are ignored.

Value

DataFrame object, whose colnames are defined using makeNames(basename(bwUrls)). Each column is type NumericList, which is a list of numeric coverage values.

Details

This function takes a GRanges object to define genomic regions, for which coverage data is loaded for each bwUrls input file.

Note that this function uses rtracklayer::import.bw() which they describe does not work on the Windows platform.

See also