Import splice junction data from BED or SJ.out.tab file

Import splice junction data from BED or SJ.out.tab file

import_juncs_from_bed(
  iBed,
  juncNames,
  sample_id,
  scale_factor,
  use_memoise = FALSE,
  memoise_junction_path = "junctions_memoise",
  gr = NULL,
  verbose = FALSE,
  ...
)

Arguments

iBed
path or URL to one BED file containing splice junction data. The score is expected to be stored in the name column (column 3), primarily because the score column is sometimes restricted to maximum value 1000. However if the name column cannot be converted to numeric without creating NA values, then the score column will be used.

juncNames
the name of the junction source file

sample_id
character string representing the sample identifier.

scale_factor
numeric value used to adjust the raw score, applied by multiplying the scale_factor by each score.

gr
GRanges representing the overall range for which junction data will be retrieved. Note that any junctions that span this range, but do not start or end inside this range, will be removed.

Details

This function is intended to be called internally by prepareSashimi(), and is provided primarily to enable use of memoise::memoise() to cache results.

This function was refactored in version 0.0.69.900 to handle either BED format, or "SJ.out.tab" junction format as produced by STAR alignment. The method uses data.table::fread() then if there are 9 columns, it assumes the format is "SJ.out.tab". Otherwise it coerces the data.frame with as(bed, "GRanges").

The BED or "SJ.out.tab" file can be gzipped, provided data.table() is able to recognieze and import the compression format.

Note that bigBed format still cannot be used since the rtracklayer package does not support that format.

Also note that junctions with score=0 are dropped at this step, to prevent propagation of junctions with zero counts.

See also