Sashimi Shiny app constants

Sashimi Shiny app constants

sashimiAppConstants(...)

Details

This function defines several constant values used by the R-shiny Splicejam Sashimi viewer, typically taking from the global environment where possible.

The R-shiny app is started by launchSashimiApp(), which calls shiny::shinyApp(), using arguments server, ui, onStart, and options. This function fulfills the argument onStart.

This function is intended to be called only from within an R-shiny app, since it defines several variables in the parent (global) environment.

The following values will be used from the environment, searching up the environment parent chain until it finds a match, until searching the global environment. Similarly, this function also defines variables in the environment using the <<- operator, which by default also searches up the environment chain until it finds a match, otherwise populating the global environment.

If a variable is not found, the corresponding default from the farrisdata package will be used.

filesDF: data.frame with colnames: "sample_id", "type", "url", "scale_factor". When "filesDF" does not exist, it is taken from farrisdata::farris_sashimi_files_df, and the default "aboutExtra" text is used.

color_sub: character vector of R colors, whose names are used to match "sample_id" or other values used to colorize ggplot visualizations. It is either populated from farrisdata::colorSub, or colorjam::group2colors() is called using the unique sample_id values in filesDF.

txdb: TxDb object used to derive exonsByTx, and cdsByTx if either does not exist. If the TxDb object does not exist, it is created from the gtf file using GenomicFeatures::makeTxDbFromGFF().

tx2geneDF: data.frame with colnames: "transcript_id", "gene_name". This data is used to convert "gene_name" to a vector of "transcript_id" values.

gtf: character path to a GTF/GFF/GFF3 file, suitable for GenomicFeatures::makeTxDbFromGFF(). The gtf is only used if tx2geneDF or exonsByTx are not available. Note that when gtf points to a remote server, the file is copied to the current working directory. If the file already exists in the local directory, it is re-used.

exonsByTx: GRangesList object, named by "transcript_id", containing all exons for each transcript.

cdsByTx: GRangesList object, named by "transcript_id", containing only CDS (protein-coding) exons for each transcript.

detectedTx: character vector of "transcript_id" values, representing the subset of transcripts detected above background. If it does not exist, it is derived from farrisdata::farrisTxSE "TxDetectedByTPM". If those values are not all present in tx2geneDF then detectedTx is defined by all transcripts present in tx2geneDF$transcript_id.

detectedGenes: inferred using tx2geneDF and detectedTx.

flatExonsByGene: GRangesList object containing non-overlapping exons for each gene, or it is derived from exonsByTx, cdsByTx, detectedTx, and tx2geneDF, using flattenExonsBy().

flatExonsByTx: GRangesList object containing non-overlapping exons for each gene, or it is derived from exonsByTx, cdsByTx, detectedTx, and tx2geneDF, using flattenExonsBy().

Several R objects are cached using memoise::memoise(), to avoid having to create the R object each time the R-shiny app is started:

flatExonsByGene

flatExonsByTx

exonsByTx

cdsByTx

A special custom variable is used to describe the specific data included with the R-shiny app, "aboutExtra". The "aboutExtra" variable is expected to contain HTML tags, for example from htmltools::tags() to format the text. The "aboutExtra" content is displayed on the tab "About Sashimi Plots" at the top.

See also